% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPACTaf.R
\docType{package}
\name{IMPACTaf}
\alias{IMPACTaf-package}
\alias{IMPACTaf}
\title{IMPACTaf: A package with functions for the IMPACTaf model.}
\description{
Description of your package.
}
\author{
Chris Kypridemos
}
